% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{barrier}
\alias{barrier}
\title{barrier}
\usage{
barrier(app_id, no_more_tasks)
}
\description{
Halt all the tasks: Notify the runtime that our current application wants to "execute" a barrier. Program will be blocked in GS_BarrierNew until all running tasks have ended. Notifies the 'no more tasks' boolean value.
}
