% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCOMPSs-package.R
\name{compss_barrier}
\alias{compss_barrier}
\title{COMPSs barrier}
\usage{
compss_barrier(app_id = 0L, no_more_tasks = FALSE)
}
\arguments{
\item{app_id}{Application ID}

\item{no_more_tasks}{Whether there are no more tasks}
}
\description{
Wrapper for barrier()
}
