% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCOMPSs-package.R
\name{compss_serialize}
\alias{compss_serialize}
\title{Serialize R object}
\usage{
compss_serialize(object, filepath, ser_method, mthreads = 1L)
}
\arguments{
\item{object}{R object to serialize}

\item{filepath}{File path to save serialized object}

\item{ser_method}{Serialization method ("cpp" preferred; legacy "qs"/"RMVL" for reading old files)}

\item{mthreads}{Number of threads for serialization}
}
\description{
Wrapper for rcompss_serialize()
}
