% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCOMPSs-package.R
\name{compss_unserialize}
\alias{compss_unserialize}
\title{Unserialize R object}
\usage{
compss_unserialize(filepath, mthreads = 1L)
}
\arguments{
\item{filepath}{File path to read serialized object from}

\item{mthreads}{Number of threads for unserialization}
}
\description{
Wrapper for rcompss_unserialize()
}
