% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCOMPSs-package.R
\name{compss_wait_on}
\alias{compss_wait_on}
\title{Wait on future object}
\usage{
compss_wait_on(future_obj, mthreads = 1L, nthreads = 1L)
}
\arguments{
\item{future_obj}{Future object to wait on}

\item{mthreads}{Number of threads for serialization}

\item{nthreads}{Number of threads for parallel processing}
}
\description{
Wrapper for rcompss_wait_on()
}
