% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCOMPSs-package.R
\name{extrae_emit_event}
\alias{extrae_emit_event}
\title{Emit EXTRAE event}
\usage{
extrae_emit_event(group, id)
}
\arguments{
\item{group}{Integer defining the event group.}

\item{id}{Integer defining the event identifier.}
}
\description{
Emit EXTRAE event
}
