% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parType_mapping}
\alias{parType_mapping}
\title{Return the type of <arg> in the COMPSs numbering system}
\usage{
parType_mapping(arg)
}
\description{
Return the type of <arg> in the COMPSs numbering system
}
