% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{process_task}
\alias{process_task}
\title{process_task}
\usage{
process_task(
  app_id,
  signatureSEXP,
  on_failureSEXP,
  time_out,
  priority,
  num_nodes,
  reduce,
  chunk_size,
  replicated,
  distributed,
  has_target,
  num_returns,
  values,
  names,
  compss_types,
  compss_directions,
  compss_streams,
  compss_prefixes,
  content_types,
  weights,
  keep_renames
)
}
\description{
Define the Rcpp function
}
