% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_context.R
\name{RCOMPSs.CopyFromGpu}
\alias{RCOMPSs.CopyFromGpu}
\title{Copy GPU buffer to R (device to host). Returns a new numeric vector.}
\usage{
RCOMPSs.CopyFromGpu(gpu_handle)
}
\arguments{
\item{gpu_handle}{Handle from RCOMPSs.GpuAlloc(n).}
}
\value{
Numeric vector of length n.
}
\description{
Copy GPU buffer to R (device to host). Returns a new numeric vector.
}
