% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_context.R
\name{RCOMPSs.CopyToGpu}
\alias{RCOMPSs.CopyToGpu}
\title{Copy R numeric vector to GPU buffer (host to device).}
\usage{
RCOMPSs.CopyToGpu(r_vec, gpu_handle)
}
\arguments{
\item{r_vec}{Numeric vector (host).}

\item{gpu_handle}{Handle from RCOMPSs.GpuAlloc(n); length(r_vec) must be <= n.}
}
\description{
Copy R numeric vector to GPU buffer (host to device).
}
