% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_context.R
\name{RCOMPSs.CreateRunContext}
\alias{RCOMPSs.CreateRunContext}
\title{Create a run context by name.}
\usage{
RCOMPSs.CreateRunContext(contextname)
}
\arguments{
\item{contextname}{Name of the new context.}
}
\description{
Create a run context by name.
}
