% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_context.R
\name{RCOMPSs.DeleteRunContext}
\alias{RCOMPSs.DeleteRunContext}
\title{Delete a run context.}
\usage{
RCOMPSs.DeleteRunContext(contextname)
}
\arguments{
\item{contextname}{Context name to delete.}
}
\description{
Delete a run context.
}
