% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_context.R
\name{RCOMPSs.FinalizeRunContext}
\alias{RCOMPSs.FinalizeRunContext}
\title{Finalize a run context (sync and free work buffers).}
\usage{
RCOMPSs.FinalizeRunContext(contextname)
}
\arguments{
\item{contextname}{Context name.}
}
\description{
Finalize a run context (sync and free work buffers).
}
