% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_context.R
\name{RCOMPSs.GetOperationPlacement}
\alias{RCOMPSs.GetOperationPlacement}
\title{Get operation placement for a context.}
\usage{
RCOMPSs.GetOperationPlacement(contextname = "default")
}
\arguments{
\item{contextname}{Context name (default "default").}
}
\description{
Get operation placement for a context.
}
