% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_context.R
\name{RCOMPSs.GetRunMode}
\alias{RCOMPSs.GetRunMode}
\title{Get run mode for a context.}
\usage{
RCOMPSs.GetRunMode(contextname = "default")
}
\arguments{
\item{contextname}{Context name (default "default").}
}
\description{
Get run mode for a context.
}
