% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_context.R
\name{RCOMPSs.GpuAlloc}
\alias{RCOMPSs.GpuAlloc}
\title{Allocate GPU buffer from R (number of double elements).
Returns an external pointer handle. Free with RCOMPSs.GpuFree or by GC.
Active context must be GPU.}
\usage{
RCOMPSs.GpuAlloc(n)
}
\arguments{
\item{n}{Number of double elements.}
}
\value{
External pointer (GPU buffer handle).
}
\description{
Allocate GPU buffer from R (number of double elements).
Returns an external pointer handle. Free with RCOMPSs.GpuFree or by GC.
Active context must be GPU.
}
