% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_context.R
\name{RCOMPSs.GpuFree}
\alias{RCOMPSs.GpuFree}
\title{Free GPU buffer. Safe to call multiple times; handle is invalidated.}
\usage{
RCOMPSs.GpuFree(gpu_handle)
}
\arguments{
\item{gpu_handle}{Handle from RCOMPSs.GpuAlloc.}
}
\description{
Free GPU buffer. Safe to call multiple times; handle is invalidated.
}
