% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_context.R
\name{RCOMPSs.SetOperationContext}
\alias{RCOMPSs.SetOperationContext}
\title{Set the active operation context.}
\usage{
RCOMPSs.SetOperationContext(contextname)
}
\arguments{
\item{contextname}{Context name to use for upcoming GPU operations.}
}
\description{
Set the active operation context.
}
