% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_context.R
\name{RCOMPSs.SetRunMode}
\alias{RCOMPSs.SetRunMode}
\title{Set run mode (sync/async) for a context.
One arg = runmode for "default"; two args = (contextname, runmode).}
\usage{
RCOMPSs.SetRunMode(contextname = "default", runmode = NULL)
}
\arguments{
\item{contextname}{Context name (default "default" when only runmode is given).}

\item{runmode}{"sync" or "async".}
}
\description{
Set run mode (sync/async) for a context.
One arg = runmode for "default"; two args = (contextname, runmode).
}
