% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_context.R
\name{RCOMPSs.SyncContext}
\alias{RCOMPSs.SyncContext}
\title{Synchronize a context (wait for its stream).}
\usage{
RCOMPSs.SyncContext(contextname)
}
\arguments{
\item{contextname}{Context name.}
}
\description{
Synchronize a context (wait for its stream).
}
