% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_build_argument_filename}
\alias{rcompss_build_argument_filename}
\title{Build serialization filename for arguments (C++ implementation)}
\usage{
rcompss_build_argument_filename(
  master_working_dir,
  ser_method,
  arg_name,
  index,
  uid
)
}
\arguments{
\item{master_working_dir}{Master working directory path}

\item{ser_method}{Serialization method (e.g., "qs", "RMVL")}

\item{arg_name}{Argument name}

\item{index}{Argument index (1-based, as R uses 1-based indexing)}

\item{uid}{Unique ID string}
}
\value{
Full filepath for serialized argument
}
\description{
This is a C++ implementation that replicates:
paste0(MASTER_WORKING_DIR, "/", ser_method, "-", arg_name, "_arg[", index, "]_", uid)
}
