% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_build_register_marker}
\alias{rcompss_build_register_marker}
\title{Build register marker name (C++ implementation)}
\usage{
rcompss_build_register_marker(function_name)
}
\arguments{
\item{function_name}{Function name}
}
\value{
Register marker string
}
\description{
This is a C++ implementation that replicates:
paste0("registered_", f_name)
}
