% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_build_return_value_filename}
\alias{rcompss_build_return_value_filename}
\title{Build serialization filename for return values (C++ implementation)}
\usage{
rcompss_build_return_value_filename(master_working_dir, ser_method, uid)
}
\arguments{
\item{master_working_dir}{Master working directory path}

\item{ser_method}{Serialization method (e.g., "qs", "RMVL")}

\item{uid}{Unique ID string}
}
\value{
Full filepath for serialized return value
}
\description{
This is a C++ implementation that replicates:
paste0(MASTER_WORKING_DIR, "/", ser_method, "-", "ReturnValue_", uid)
}
