% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_build_return_value_name}
\alias{rcompss_build_return_value_name}
\title{Build return value name (C++ implementation)}
\usage{
rcompss_build_return_value_name(ser_method)
}
\arguments{
\item{ser_method}{Serialization method (e.g., "qs", "RMVL")}
}
\value{
Return value name string
}
\description{
This is a C++ implementation that replicates:
paste0(ser_method, "-RETURN_VALUE")
}
