% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_build_type_args_path}
\alias{rcompss_build_type_args_path}
\title{Build typeArgs path (C++ implementation)}
\usage{
rcompss_build_type_args_path(working_dir, filename)
}
\arguments{
\item{working_dir}{Current working directory}

\item{filename}{Filename}
}
\value{
Full path string
}
\description{
This is a C++ implementation that replicates:
paste0(getwd(), "/", filename)
}
