% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_build_variable_argument_names}
\alias{rcompss_build_variable_argument_names}
\title{Build variable argument names (C++ implementation)}
\usage{
rcompss_build_variable_argument_names(function_name, count)
}
\arguments{
\item{function_name}{Function name}

\item{count}{Number of arguments}
}
\value{
Character vector of argument names
}
\description{
This is a C++ implementation that replicates:
paste0(paste0(f_name, "___"), 1:length(arguments))
Generates names like: "functionName___1", "functionName___2", etc.
}
