% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_concat_integer}
\alias{rcompss_concat_integer}
\title{Concatenate integer vector with value (C++ implementation)}
\usage{
rcompss_concat_integer(vec, value)
}
\arguments{
\item{vec}{Input integer vector}

\item{value}{Value to append}
}
\value{
New integer vector with value appended
}
\description{
This is a C++ implementation that replicates: c(vector, value)
}
