% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_concat_string}
\alias{rcompss_concat_string}
\title{Concatenate string vector with value (C++ implementation)}
\usage{
rcompss_concat_string(vec, value)
}
\arguments{
\item{vec}{Input character vector}

\item{value}{Value to append}
}
\value{
New character vector with value appended
}
\description{
This is a C++ implementation that replicates: c(vector, value)
}
