% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_copy_from_gpu}
\alias{rcompss_copy_from_gpu}
\title{Copy GPU buffer to R (device to host). Returns a new numeric vector.}
\usage{
rcompss_copy_from_gpu(gpu_handle)
}
\description{
Copy GPU buffer to R (device to host). Returns a new numeric vector.
}
