% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_copy_to_gpu}
\alias{rcompss_copy_to_gpu}
\title{Copy R numeric vector to GPU buffer (host to device).}
\usage{
rcompss_copy_to_gpu(r_vec, gpu_handle)
}
\description{
Copy R numeric vector to GPU buffer (host to device).
}
