% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_extract_serialization_method}
\alias{rcompss_extract_serialization_method}
\title{Extract serialization method from filepath (C++ implementation)}
\usage{
rcompss_extract_serialization_method(filepath)
}
\arguments{
\item{filepath}{Full file path}
}
\value{
Serialization method string (e.g., "qs", "RMVL")
}
\description{
This is a C++ implementation that replicates: strsplit(basename(filepath), "-")[[1]][1]
Extracts the method prefix from filenames like: /path/to/qs-filenamexxxx
}
