% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_generate_uid}
\alias{rcompss_generate_uid}
\title{Generate a unique ID (C++ implementation)}
\usage{
rcompss_generate_uid()
}
\value{
A unique string ID (format: YYYYMMDDHHMMSS-randomstring)
}
\description{
This is a C++ implementation that exactly replicates the R UID() function logic.
Uses the UIDGenerator class which replicates:
  current_time <- Sys.time()
  time_string <- format(current_time, "%Y%m%d%H%M%S")
  random_string <- paste0(sample(c(letters, LETTERS, 0:9), 50, replace = TRUE), collapse = "")
  return(paste0(time_string, "-", random_string))
}
