% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_gpu_daxpy}
\alias{rcompss_gpu_daxpy}
\title{y <- alpha*x + y on GPU (cuBLAS DAXPY). Returns a new numeric vector (does not modify x).
Requires an active GPU operation context (set from R before calling).}
\usage{
rcompss_gpu_daxpy(x, y, alpha = 1)
}
\description{
y <- alpha*x + y on GPU (cuBLAS DAXPY). Returns a new numeric vector (does not modify x).
Requires an active GPU operation context (set from R before calling).
}
