% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_gpu_dscal}
\alias{rcompss_gpu_dscal}
\title{Scale a vector on GPU (cuBLAS DSCAL): returns alpha * x (new vector).}
\usage{
rcompss_gpu_dscal(x, alpha = 1)
}
\description{
Scale a vector on GPU (cuBLAS DSCAL): returns alpha * x (new vector).
}
