% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_gpu_dsyrk}
\alias{rcompss_gpu_dsyrk}
\title{Symmetric rank-k update on GPU (cuBLAS DSRYK).
C = alpha * op(A) %*% t(op(A)) + beta * C, with C symmetric.
Only the triangle specified by \code{uplo} is referenced/updated.}
\usage{
rcompss_gpu_dsyrk(A, uplo = "U", trans = "N", alpha = 1, beta = 0, C0 = NULL)
}
\description{
Symmetric rank-k update on GPU (cuBLAS DSRYK).
C = alpha * op(A) %*% t(op(A)) + beta * C, with C symmetric.
Only the triangle specified by \code{uplo} is referenced/updated.
}
