% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_gpu_free}
\alias{rcompss_gpu_free}
\title{Free GPU buffer. Safe to call multiple times; handle is invalidated.}
\usage{
rcompss_gpu_free(gpu_handle)
}
\description{
Free GPU buffer. Safe to call multiple times; handle is invalidated.
}
