% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_gpu_vector_add_kernel}
\alias{rcompss_gpu_vector_add_kernel}
\title{Run vector addition kernel only (inputs and output are GPU buffers allocated from R).}
\usage{
rcompss_gpu_vector_add_kernel(d_a_handle, d_b_handle, d_result_handle)
}
\description{
Run vector addition kernel only (inputs and output are GPU buffers allocated from R).
}
