% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_is_basic_type}
\alias{rcompss_is_basic_type}
\title{Check if R object is a basic type (C++ implementation)}
\usage{
rcompss_is_basic_type(arg)
}
\arguments{
\item{arg}{R object to check}
}
\value{
true if object is a basic type, false otherwise
}
\description{
This is a C++ implementation that replicates:
length(class(arg)) == 1 && length(arg) == 1 && (class(arg) %in% c("integer", "numeric", "character"))
}
