% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_is_future_object_path}
\alias{rcompss_is_future_object_path}
\title{Check if R object is a future_object_path (C++ implementation)}
\usage{
rcompss_is_future_object_path(obj)
}
\arguments{
\item{obj}{R object to check}
}
\value{
true if object is a future_object_path, false otherwise
}
\description{
This is a C++ implementation that replicates:
length(class(obj)) == 1 && class(obj) == "future_object_path"
}
