% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_is_list}
\alias{rcompss_is_list}
\title{Check if R object is a list (C++ implementation)}
\usage{
rcompss_is_list(obj)
}
\arguments{
\item{obj}{R object to check}
}
\value{
true if object is a list, false otherwise
}
\description{
This is a C++ implementation that replicates:
length(class(obj)) == 1 && class(obj) == "list"
}
