% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_par_type_mapping}
\alias{rcompss_par_type_mapping}
\title{Map R type to COMPSs type number (C++ implementation)}
\usage{
rcompss_par_type_mapping(arg)
}
\arguments{
\item{arg}{R object to get type from}
}
\value{
COMPSs type number (integer)
}
\description{
This is a C++ implementation that exactly replicates the R parType_mapping() function logic.
Maps R types to COMPSs numbering system:
  "logical" = 0L
  "CHAR" = 1L
  "BYTE" = 2L
  "SHORT" = 3L
  "integer" = 4L
  "double" = 7L
  "character" = 8L
  default = 10L (FILE)
}
