% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_rep_integer}
\alias{rcompss_rep_integer}
\title{Create integer vector filled with value (C++ implementation)}
\usage{
rcompss_rep_integer(value, count)
}
\arguments{
\item{value}{Integer value to repeat}

\item{count}{Length of vector}
}
\value{
Integer vector filled with value
}
\description{
This is a C++ implementation that replicates: rep(value, length = count)
}
