% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_rep_string}
\alias{rcompss_rep_string}
\title{Create string vector filled with value (C++ implementation)}
\usage{
rcompss_rep_string(value, count)
}
\arguments{
\item{value}{String value to repeat}

\item{count}{Length of vector}
}
\value{
Character vector filled with value
}
\description{
This is a C++ implementation that replicates: rep(value, length = count)
}
