% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcompss_wait_on}
\alias{rcompss_wait_on}
\title{Wait on future object (C++ implementation)}
\usage{
rcompss_wait_on(future_obj, mthreads = 1L, nthreads = 1L)
}
\arguments{
\item{future_obj}{Future object to wait on}

\item{mthreads}{Number of threads for serialization}

\item{nthreads}{Number of threads for parallel processing}
}
\value{
Deserialized result
}
\description{
This C++ function handles waiting on future objects, lists, and future_object_path.
}
