% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{register_core_element}
\alias{register_core_element}
\title{Register a core element}
\usage{
register_core_element(
  CESignature,
  ImplSignature,
  ImplConstraints,
  ImplType,
  ImplLocal,
  ImplIO,
  prolog,
  epilog,
  container,
  typeArgs
)
}
\description{
Register a core element
}
