% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{start_runtime_interactive}
\alias{start_runtime_interactive}
\title{Start a COMPSs-Runtime instance within interactive session}
\usage{
start_runtime_interactive(debug = FALSE, graph = FALSE, trace = FALSE)
}
\description{
Note: Many parameters are currently commented out but may be added in the future:
monitor, project_xml, resources_xml, summary, task_execution, storage_conf, etc.
}
