% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCOMPSs-package.R, R/RcppExports.R
\name{task}
\alias{task}
\title{Task decorator}
\usage{
task(
  f,
  filename,
  return_value = FALSE,
  return_type = "list",
  ser_method = as.character(c("cpp", "cpp")),
  info_only = FALSE,
  DEBUG = FALSE,
  f_name = ""
)

task(
  f,
  filename,
  return_value = FALSE,
  return_type = "list",
  ser_method = as.character(c("cpp", "cpp")),
  info_only = FALSE,
  DEBUG = FALSE,
  f_name = ""
)
}
\arguments{
\item{f}{The function to be executed}

\item{filename}{Character. The file where the function is defined}

\item{return_value}{Boolean. Default value is FALSE. Whether there is a return value}

\item{return_type}{Character. Default value is "list". Return type ("list" or "element")}

\item{ser_method}{Character vector. Default value is c("qs", "qs"). Serialization method}

\item{info_only}{Boolean. Whether the run is to print the information only}

\item{DEBUG}{Boolean. Whether to print debug information}

\item{f_name}{Character. The name of the function (extracted from match.call() in R wrapper)}
}
\value{
The decorated function (R function closure)
}
\description{
Main task decorator function. The actual implementation is in C++ TaskDecorator class.

This is the main entry point that creates the task decorator.
All logic is in C++ TaskDecorator class - this is just a thin Rcpp export wrapper.
}
