% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCOMPSs-package.R
\name{UID}
\alias{UID}
\title{Generate a unique ID}
\usage{
UID()
}
\value{
A unique string ID (format: YYYYMMDDHHMMSS-randomstring)
}
\description{
This function generates a unique identifier using a C++ implementation.
The UID format is: YYYYMMDDHHMMSS-randomstring
}
