% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compss_barrier}
\alias{compss_barrier}
\title{compss_barrier}
\usage{
compss_barrier(no_more_tasks = FALSE)
}
\arguments{
\item{no_more_tasks}{Boolean.}
}
\description{
Barrier for the COMPSs runtime system
}
