% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compss_wait_on}
\alias{compss_wait_on}
\title{compss_wait_on}
\usage{
compss_wait_on(future_obj)
}
\arguments{
\item{future_obj}{}
}
\description{
Serialization in R and synchronize the results with the master
}
