% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{task}
\alias{task}
\title{task}
\usage{
task(f, info_only = FALSE, ...)
}
\arguments{
\item{f}{The function to be executed.}

\item{...}{Metadata.}
}
\value{
The decorated function
}
\description{
This function is the decorator for tasks.
}
